Refactoring to have namespace packages for BehaviorData
