+++ LibSaberPatch/AssetData.cs
@@ -1,3 +1,4 @@
+using LibSaberPatch.BehaviorDataObjects;
 using System;
 using System.Collections.Generic;
 using System.IO;
