+++ LibSaberPatch/JSONTypes.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+using LibSaberPatch.BehaviorDataObjects;
 
 namespace LibSaberPatch
 {
