+++ LibSaberPatch/Utils.cs
@@ -6,6 +6,7 @@ using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Advanced;
 using SixLabors.ImageSharp.Processing;
 using SixLabors.ImageSharp.PixelFormats;
+using LibSaberPatch.BehaviorDataObjects;
 
 namespace LibSaberPatch
 {
